:recycle: 자식 뷰 폴더명 변경 Components -> Views, Home 뷰들 파일 분리
